+++ src/CM/Messaging/Http/ApiHttpClient.php
@@ -58,9 +58,12 @@ class ApiHttpClient
 
     /**
      * @param string $endpoint
-     * @param string $body
+     * @param array  $body
      *
      * @return Response
+     * @throws \Http\Client\Exception
+     * @throws \Exception
+     * @throws \RuntimeException
      * @throws BadRequestException
      */
     public function postRequest($endpoint, $body)