+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [1.1.0] - 2017-09-21
+### Added
+- add custom grouping support 
+
 ## [1.0.2] - 2017-06-29
 ### Fixed
 - tests
@@ -36,7 +40,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - initial code
 
-[Unreleased]: https://github.com/CMTelecom/messaging-php/compare/1.0.2....HEAD
+[Unreleased]: https://github.com/CMTelecom/messaging-php/compare/1.1.0....HEAD
+[1.1.0]: https://github.com/CMTelecom/messaging-php/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/CMTelecom/messaging-php/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/CMTelecom/messaging-php/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/CMTelecom/messaging-php/compare/0.0.1...1.0.0